Add backup.

+++ PKGBUILD
@@ -25,10 +25,11 @@ sha256sums=(
 )
 makedepends=(python-build python-installer python-wheel)
 install=$pkgname.install
+backup=("etc/ddns/config.json")
 
 package() {
 	python -m installer --destdir="$pkgdir" *.whl
         install -Dm644 newfuture_ddns.service  "${pkgdir}/usr/lib/systemd/system/newfuture_ddns.service"
         install -Dm644 newfuture_ddns.timer "${pkgdir}/usr/lib/systemd/system/newfuture_ddns.timer";
-        install -Dm644 config.json "${pkgdir}/etc/ddns/config.json";
+        [[ ! -f /etc/ddns/config.json ]] || install -Dm644 config.json "${pkgdir}/etc/ddns/config.json";
 }
